+++ lib/lpfuart/src/parseinfostate.cpp
@@ -13,7 +13,7 @@ ParseInfoState::ParseInfoState(LegoDevice *legoDevice, int messageType, int mess
 	, messageMode(messageMode)
 	, messageSize(messageSize + 1)
 	, messageType(messageType) {
-	messagePayload = new int[this->messageSize];
+	messagePayload = new uint8_t[this->messageSize];
 	received = 0;
 }
 
@@ -194,7 +194,7 @@ void ParseInfoState::parseINFO() {
 
 ProtocolState *ParseInfoState::parse(int datapoint) {
 	if (received < messageSize) {
-		messagePayload[received++] = datapoint;
+		messagePayload[received++] = static_cast<uint8_t>(datapoint);
 		return this;
 	}
